agg related code clean
